Add documentation all classes in the parser
Replace single thread parser with multi thread capability.
