+++ ada-parser/src/main/java/com/ucl/ADA/parser/parser/ADAParser.java
@@ -6,29 +6,53 @@ import com.ucl.ADA.parser.ada_model.ADAClass;
 import com.ucl.ADA.parser.parser.visitor.ADAClassVisitor;
 import com.ucl.ADA.parser.parser.visitor.PackageAndImportVisitor;
 import com.ucl.ADA.parser.parser.visitor.TypeDeclarationVisitor;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.filefilter.TrueFileFilter;
-import org.eclipse.jdt.core.JavaCore;
-import org.eclipse.jdt.core.dom.*;
+import org.apache.commons.collections4.ListUtils;
+import org.eclipse.jdt.core.dom.AbstractTypeDeclaration;
+import org.eclipse.jdt.core.dom.CompilationUnit;
 
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
+import java.io.File;
 import java.util.*;
-
+import java.util.concurrent.*;
 
 public class ADAParser {
 
-    public Set<ADAClass> getParsedSourceFile(String sourceRoot) {
+    private final int NUMBER_OF_THREADS = 8;
+    private final int NUMBER_OF_FILES_IN_A_BATCH = 80;
+    private final SourceFileProcessor sourceFileProcessor;
+    private final CompilationUnitBuilder compilationUnitBuilder;
+
+
+    /**
+     * Constructor of ADAParser
+     */
+    public ADAParser() {
+        sourceFileProcessor = new SourceFileProcessor();
+        compilationUnitBuilder = new CompilationUnitBuilder();
+    }
+
+
+    /**
+     * It parses all the .*java source files and populates a set of ADAClass model.
+     *
+     * @param rootDirectory :  Source repository path
+     * @return A set of parsed class in from of ADAClass model
+     */
+    public Set<ADAClass> getParsedSourceFile(String rootDirectory) {
         Set<ADAClass> allParsedFile = new HashSet<>();
-        List<ADAClass> classes = parseEachSourceFile(sourceRoot);
+        List<ADAClass> classes = parsingSourceFilesConcurrently(rootDirectory);
         allParsedFile.addAll(classes);
         return allParsedFile;
     }
 
-    public void printParsedSourceFileInJSON(String sourceRoot) {
-        List<ADAClass> classes = parseEachSourceFile(sourceRoot);
+
+    /**
+     * It parses all the .*java source files and populates a set of ADAClass model.
+     * It then prints the parsed ADAClasses in JSON format.
+     *
+     * @param rootDirectory :Source repository path
+     */
+    public void printParsedSourceFileInJSON(String rootDirectory) {
+        Set<ADAClass> classes = getParsedSourceFile(rootDirectory);
         for (ADAClass aClass : classes) {
             ObjectMapper objMapper = new ObjectMapper();
             String jsonStr = "[]";
@@ -41,86 +65,73 @@ public class ADAParser {
         }
     }
 
-    private List<ADAClass> parseSourceFile(ASTParser parser, String sourceFile) {
-        String sourceCode = getSourceCodeFromSourcePath(sourceFile);
-        String unitName = new File(sourceFile).getName();
-        parser.setSource(sourceCode.toCharArray());
-        parser.setUnitName(unitName);
-        CompilationUnit cu = (CompilationUnit) parser.createAST(null);
-        if (cu.getAST().hasBindingsRecovery()) {
-            //System.out.println("Binding activated.");
-        }
 
-        PackageAndImportVisitor pv = new PackageAndImportVisitor();
-        cu.accept(pv);
-        TypeDeclarationVisitor tv = new TypeDeclarationVisitor();
-        cu.accept(tv);
-
-        List<AbstractTypeDeclaration> allClassAndEnums = tv.getAbstractTypeDeclaration();
+    /**
+     * This method parsed all source *.java files for a given source repository and make a list of ADAClass models
+     * It performs thread pooling to parsing the source files and populate list of ADAClass models concurrently.
+     *
+     * @param rootDirectory: :Source repository path
+     * @return A list of parsed class in from of ADAClass model
+     */
+    private List<ADAClass> parsingSourceFilesConcurrently(String rootDirectory) {
         List<ADAClass> parsedClasses = new ArrayList<>();
-        for (AbstractTypeDeclaration cl : allClassAndEnums) {
-            ADAClassVisitor jp = new ADAClassVisitor(pv.getPackageName(), pv.getImportedInternalClasses(), pv.getImportedExternalClasses());
-            cl.accept(jp);
-            ADAClass cm = jp.getExtractedClass();
-            parsedClasses.add(cm);
+        List<String> filePaths = sourceFileProcessor.getSourceFilePaths(rootDirectory);
+        List<List<String>> filePathBatches = ListUtils.partition(filePaths, NUMBER_OF_FILES_IN_A_BATCH);
+        List<Map<String, String>> allFileContents = sourceFileProcessor.getSourceContentsInChunks(filePathBatches);
+        String[] allSrcDirectories = sourceFileProcessor.getSourceDirectories(new File(rootDirectory));
+        ExecutorService executor = Executors.newFixedThreadPool(NUMBER_OF_THREADS);
+        // Callable, return a future, submit and run the task async
+        List<Callable<List<ADAClass>>> listOfCallable = new ArrayList<>();
+        for (Map<String, String> filePathBatch : allFileContents) {
+            Callable<List<ADAClass>> c = () -> {
+                return parseSourceFiles(filePathBatch, allSrcDirectories);
+            };
+            listOfCallable.add(c);
         }
-        return parsedClasses;
-
-    }
-
-    private List<ADAClass> parseAllSourceFiles(String sourceRoot) {
-        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
-        LocalDateTime now = LocalDateTime.now();
-        System.out.println("Source files path  collection started on...." + dtf.format(now));
-        List<String> filePaths = getJavaSourceFilePathsFromSourceRoot(sourceRoot);
-        System.out.println("Collection is completed on: " + dtf.format(now));
-
-        System.out.println("Source file reading is started on......" + dtf.format(now));
-        Map<String, String> fileContents = getAllSourceContent(filePaths);
-        System.out.println("Reading is completed on: " + dtf.format(now));
-
-        System.out.println("Compilation unit preparation is started on....");
-        List<CompilationUnit> compilationUnits = getCompilationUnits(fileContents, sourceRoot);
-        System.out.println("Compilation Unit building finished: " + dtf.format(now));
-
-        System.out.println("ADA parser is started on.........." + dtf.format(now));
-        List<ADAClass> parsedClasses = new ArrayList<>();
-        for (CompilationUnit cu : compilationUnits) {
-            List<ADAClass> classes = getParsedClass(cu);
-            parsedClasses.addAll(classes);
+        try {
+            List<Future<List<ADAClass>>> futures = executor.invokeAll(listOfCallable);
+            for (Future<List<ADAClass>> future : futures) {
+                if (future.isDone()) {
+                    parsedClasses.addAll(future.get());
+                }
+            }
+        } catch (InterruptedException | ExecutionException e) {
+            // thread was interrupted
+            e.printStackTrace();
+        } finally {
+            // shut down the executor manually
+            executor.shutdown();
         }
-        System.out.println("ADAClass Model building finished on: " + dtf.format(now));
         return parsedClasses;
     }
 
-    private Map<String, String> getFileContents(String sourceRoot) {
-        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
-        LocalDateTime now = LocalDateTime.now();
-        System.out.println("Source files path  collection started on...." + dtf.format(now));
-        List<String> filePaths = getJavaSourceFilePathsFromSourceRoot(sourceRoot);
-        System.out.println("Collection is completed on: " + dtf.format(now));
-        System.out.println("Source file reading is started on......" + dtf.format(now));
-        Map<String, String> fileContents = new HashMap<>();
-        fileContents.putAll(getAllSourceContent(filePaths));
-        System.out.println("Reading is completed on: " + dtf.format(now));
-        System.out.println("ADA parser is started on.........." + dtf.format(now));
-        return fileContents;
-    }
 
-    private List<ADAClass> parseEachSourceFile(String sourceRoot) {
-        Map<String, String> fileContents = getFileContents(sourceRoot);
+    /**
+     * It parses a given list of files and populates a list of ADAClass model.
+     *
+     * @param filePathContentPair : A map containing file path->file contents
+     * @param sourceDirectories   : All the source directories (src/) insider the repository.
+     * @return A list of ADAClass model with parsed data.
+     */
+    private List<ADAClass> parseSourceFiles(Map<String, String> filePathContentPair, String[] sourceDirectories) {
         List<ADAClass> parsedClasses = new ArrayList<>();
-        String[] allSrcDirectories = getSourceDirectories(new File(sourceRoot));
-        for (Map.Entry<String, String> sourceEntry : fileContents.entrySet()) {
+        for (Map.Entry<String, String> sourceEntry : filePathContentPair.entrySet()) {
             String filePath = sourceEntry.getKey();
             String sourceCode = sourceEntry.getValue();
-            CompilationUnit compilationUnit = getCompilationUnit(filePath, sourceCode, allSrcDirectories);
+            CompilationUnit compilationUnit = compilationUnitBuilder.getCompilationUnit(filePath, sourceCode, sourceDirectories);
             List<ADAClass> classes = getParsedClass(compilationUnit);
             parsedClasses.addAll(classes);
         }
         return parsedClasses;
     }
 
+
+    /**
+     * It takes a prepared CompilationUnit and generates a list of parsed ADAClass model.
+     *
+     * @param compilationUnit : A compilation unit derived from CompilationUnit builder.
+     * @return A list of parsed ADAClass model.
+     */
     private List<ADAClass> getParsedClass(CompilationUnit compilationUnit) {
         List<ADAClass> parsedClasses = new ArrayList<>();
         if (compilationUnit != null) {
@@ -142,122 +153,5 @@ public class ADAParser {
         return parsedClasses;
     }
 
-    private Map getParserVersion_1_5() {
-        Map options = JavaCore.getOptions();
-        options.put(JavaCore.COMPILER_COMPLIANCE, JavaCore.VERSION_1_5);
-        options.put(JavaCore.COMPILER_CODEGEN_TARGET_PLATFORM, JavaCore.VERSION_1_5);
-        options.put(JavaCore.COMPILER_SOURCE, JavaCore.VERSION_1_5);
-        return options;
-    }
-
-    private ASTParser buildASTParser(String[] sourceDirectories, Map options) {
-        String[] encoding = new String[sourceDirectories.length];
-        Arrays.fill(encoding, "UTF-8");
-        ASTParser parser = ASTParser.newParser(AST.JLS13);
-        parser.setKind(ASTParser.K_COMPILATION_UNIT);
-        parser.setBindingsRecovery(true);
-        parser.setResolveBindings(true);
-        parser.setStatementsRecovery(true);
-        parser.setCompilerOptions(options);
-        String[] classpath = {};
-        parser.setEnvironment(classpath, sourceDirectories, encoding, true);
-        return parser;
-    }
-
-    private CompilationUnit getCompilationUnit(String filePath, String sourceCodes, String[] allSrcDirectories) {
-        Map options = getParserVersion_1_5();
-        ASTParser parser = buildASTParser(allSrcDirectories, options);
-        String unitName = new File(filePath).getName();
-        parser.setSource(sourceCodes.toCharArray());
-        parser.setUnitName(unitName);
-        CompilationUnit compilationUnit = null;
-        try {
-            compilationUnit = (CompilationUnit) parser.createAST(null);
-        } catch (Exception ex) {
-            System.err.println("Parsing Error at file-> " + filePath);
-        }
-        return compilationUnit;
-    }
-
-    private List<String> getJavaSourceFilePathsFromSourceRoot(String rootPath) {
-        String[] allSrcDirectories = getSourceDirectories(new File(rootPath));
-        List<String> sourceFiles = new ArrayList<>();
-        final String[] suffix = {"java"};
-        for (String srcDir : allSrcDirectories) {
-            File rootDir = new File(srcDir);
-            Collection<File> files = FileUtils.listFiles(rootDir, suffix, true);
-            for (File file : files) {
-                if (!file.getPath().endsWith("Test.java") && !file.getPath().endsWith("Tests.java")) {
-                    {
-                        sourceFiles.add(file.getAbsolutePath());
-                    }
-                }
-            }
-        }
-        return sourceFiles;
-    }
-
-    private String[] getSourceDirectories(File rootDir) {
-        List<String> allSourceDirectories = new ArrayList<>();
-        Collection<File> files = FileUtils.listFilesAndDirs(rootDir,
-                TrueFileFilter.INSTANCE,
-                TrueFileFilter.INSTANCE);
-        for (File file : files) {
-            if (file.isDirectory()) {
-                if (file.getName().equals("src") || file.getName().equals("Src"))
-                    allSourceDirectories.add(file.getAbsolutePath());
-            }
-        }
-        String[] srcDirs = allSourceDirectories.stream().toArray(String[]::new);
-        return srcDirs;
-    }
-
-    private String getSourceCodeFromSourcePath(String filePath) {
-        String sourceCode = "";
-        File sourceFile = new File(filePath);
-        if (sourceFile.exists()) {
-            try {
-                sourceCode = FileUtils.readFileToString(sourceFile, StandardCharsets.UTF_8);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-        return sourceCode;
-    }
-
-    private Map<String, String> getAllSourceContent(List<String> paths) {
-        Map<String, String> sourceContents = new HashMap<>();
-        for (String path : paths) {
-            String code = getSourceCodeFromSourcePath(path);
-            sourceContents.put(path, code);
-        }
-        return sourceContents;
-    }
-
-    private List<CompilationUnit> getCompilationUnits(Map<String, String> sourceCodes, String sourceRoot) {
-        String[] allSrcDirectories = getSourceDirectories(new File(sourceRoot));
-        List<CompilationUnit> compilationUnits = new ArrayList<>();
-        int count = 0;
-        for (String path : sourceCodes.keySet()) {
-            try {
-                count++;
-                Map options = getParserVersion_1_5();
-                ASTParser parser = buildASTParser(allSrcDirectories, options);
-                String code = sourceCodes.get(path);
-                String unitName = new File(path).getName();
-                parser.setSource(code.toCharArray());
-                parser.setUnitName(unitName);
-                System.out.println(count + "->" + path);
-                CompilationUnit cu = (CompilationUnit) parser.createAST(null);
-                compilationUnits.add(cu);
-            } catch (Exception e) {
-                System.err.println("Exception occurred in " + path);
-            }
-        }
-        return compilationUnits;
-    }
-
 
 }
-
-
