+++ ada-parser/src/main/java/com/ucl/ADA/parser/parser/visitor/PackageAndImportVisitor.java
@@ -1,38 +1,37 @@
 package com.ucl.ADA.parser.parser.visitor;
 
+import lombok.Getter;
 import org.eclipse.jdt.core.dom.ASTVisitor;
 import org.eclipse.jdt.core.dom.IBinding;
 import org.eclipse.jdt.core.dom.ImportDeclaration;
 import org.eclipse.jdt.core.dom.PackageDeclaration;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
+@Getter
 public class PackageAndImportVisitor extends ASTVisitor {
 
-    String packageName = "";
-    Set<String> importedInternalClasses;
-    Set<String> importedExternalClasses;
+    private String packageName = "";
+    private Set<String> importedInternalClasses;
+    private Set<String> importedExternalClasses;
+
+    /**
+     * A constructor of PackageAndImportVisitor.
+     */
 
     public PackageAndImportVisitor() {
         importedInternalClasses = new HashSet<>();
         importedExternalClasses = new HashSet<>();
     }
 
-    public String getPackageName() {
-        return packageName;
-    }
-
-    public Set<String> getImportedInternalClasses() {
-        return importedInternalClasses;
-    }
-
-    public Set<String> getImportedExternalClasses() {
-        return importedExternalClasses;
-    }
-
-    // package declaration
+    /**
+     * It visits the PackageDeclaration node from the AST and
+     * retrieves the declared package name
+     *
+     * @param node : A PackageDeclaration node derived from the AST.
+     * @return true if it is required to visit the children node otherwise false
+     */
     public boolean visit(PackageDeclaration node) {
         IBinding binding = node.resolveBinding();
         if (binding != null) {
@@ -41,7 +40,13 @@ public class PackageAndImportVisitor extends ASTVisitor {
         return true;
     }
 
-    // import declaration
+    /**
+     * It visits the ImportDeclaration node from the AST and
+     * retrieves the required imported packages names.
+     *
+     * @param node : A ImportDeclaration node derived from the AST.
+     * @return true if it is required to visit the children node otherwise false
+     */
     public boolean visit(ImportDeclaration node) {
         IBinding binding = node.resolveBinding();
         if (binding != null) {
