+++ ada-parser/src/main/java/com/ucl/ADA/parser/transformer/Transformer.java
@@ -3,7 +3,6 @@ package com.ucl.ADA.parser.transformer;
 import com.ucl.ADA.model.project_structure.ProjectStructure;
 import com.ucl.ADA.parser.ada_model.ADAClass;
 import com.ucl.ADA.parser.parser.ADAParser;
-import com.ucl.ADA.parser.parser.ExecutorParser;
 
 import java.util.Set;
 
@@ -39,12 +38,11 @@ public class Transformer {
         //String src_dir = "ada-parser/src/main/resources/source_to_parse";
         //String src_dir ="/home/mrhmisu/Downloads/hadoop-trunk/";
         //String src_dir = "/home/mrhmisu/Downloads/mockito-release-3.x/";
-        //String src_dir = "/home/mrhmisu/UCL-MS/ADA-test-simple-JAVA-project-0/src";
+        String src_dir = "/home/mrhmisu/UCL-MS/ADA-test-simple-JAVA-project-0/src";
         // String src_dir = "/home/mrhmisu/Downloads/guava-master/";
-        String src_dir = "/home/mrhmisu/UCL-MS/ADA";
+        //String src_dir = "/home/mrhmisu/UCL-MS/ADA";
         // new ADAParser().printParsedSourceFileInJSON(src_dir);
-
-        new ExecutorParser().printParsedSourceFileInJSON(src_dir);
+        new ADAParser().printParsedSourceFileInJSON(src_dir);
 
 
     }
